+++ mypyc/ir/class_ir.py
@@ -5,7 +5,7 @@ from collections import OrderedDict
 
 from mypyc.common import JsonDict
 from mypyc.ir.ops import Value, DeserMaps
-from mypyc.rtypes import RType, RInstance, deserialize_type
+from mypyc.ir.rtypes import RType, RInstance, deserialize_type
 from mypyc.ir.func_ir import FuncIR, FuncDecl, FuncSignature
 from mypyc.namegen import NameGenerator, exported_name
 from mypyc.common import PROPSET_PREFIX
