+++ mypyc/ir/module_ir.py
@@ -4,7 +4,7 @@ from typing import List, Tuple, Dict
 
 from mypyc.common import JsonDict
 from mypyc.ir.ops import DeserMaps
-from mypyc.rtypes import RType, deserialize_type
+from mypyc.ir.rtypes import RType, deserialize_type
 from mypyc.ir.func_ir import FuncIR, FuncDecl, format_func
 from mypyc.ir.class_ir import ClassIR
