+++ mypyc/irbuild/for_helpers.py
@@ -12,7 +12,7 @@ from mypy.nodes import Lvalue, Expression
 from mypyc.ir.ops import (
     Value, BasicBlock, LoadInt, PrimitiveOp, Branch, Register, AssignmentTarget
 )
-from mypyc.rtypes import RType, is_short_int_rprimitive
+from mypyc.ir.rtypes import RType, is_short_int_rprimitive
 from mypyc.primitives.int_ops import unsafe_short_add
 from mypyc.primitives.list_ops import list_len_op, list_get_item_unsafe_op
 from mypyc.primitives.misc_ops import iter_op, next_op
