+++ mypyc/irbuild/mapper.py
@@ -9,7 +9,7 @@ from mypy.types import (
 )
 
 from mypyc.ir.ops import LiteralsMap
-from mypyc.rtypes import (
+from mypyc.ir.rtypes import (
     RType, RUnion, RTuple, RInstance, object_rprimitive, dict_rprimitive, tuple_rprimitive,
     none_rprimitive, int_rprimitive, float_rprimitive, str_rprimitive, bool_rprimitive,
     list_rprimitive, set_rprimitive
