+++ mypyc/irbuild/prepare.py
@@ -8,7 +8,7 @@ from mypy.types import Type
 from mypy.build import Graph
 
 from mypyc.ir.ops import DeserMaps
-from mypyc.rtypes import RInstance, tuple_rprimitive, dict_rprimitive
+from mypyc.ir.rtypes import RInstance, tuple_rprimitive, dict_rprimitive
 from mypyc.ir.func_ir import (
     FuncDecl, FuncSignature, RuntimeArg, FUNC_NORMAL, FUNC_STATICMETHOD, FUNC_CLASSMETHOD
 )
