+++ mypyc/irbuild/specialize.py
@@ -20,7 +20,7 @@ from mypy.types import AnyType, TypeOfAny
 from mypyc.ir.ops import (
     Value, BasicBlock, LoadInt, RaiseStandardError, Unreachable, OpDescription
 )
-from mypyc.rtypes import (
+from mypyc.ir.rtypes import (
     RType, RTuple, str_rprimitive, list_rprimitive, dict_rprimitive, set_rprimitive,
     bool_rprimitive
 )
