+++ mypyc/primitives/list_ops.py
@@ -3,7 +3,7 @@
 from typing import List
 
 from mypyc.ir.ops import ERR_MAGIC, ERR_NEVER, ERR_FALSE, EmitterInterface
-from mypyc.rtypes import (
+from mypyc.ir.rtypes import (
     int_rprimitive, short_int_rprimitive, list_rprimitive, object_rprimitive, bool_rprimitive
 )
 from mypyc.primitives.registry import (
