+++ mypyc/primitives/misc_ops.py
@@ -2,7 +2,7 @@
 
 
 from mypyc.ir.ops import ERR_NEVER, ERR_MAGIC, ERR_FALSE
-from mypyc.rtypes import (
+from mypyc.ir.rtypes import (
     RTuple, none_rprimitive, bool_rprimitive, object_rprimitive, str_rprimitive,
     int_rprimitive, dict_rprimitive
 )
