+++ mypyc/primitives/registry.py
@@ -11,7 +11,7 @@ from typing import Dict, List, Optional
 from mypyc.ir.ops import (
     OpDescription, EmitterInterface, EmitCallback, StealsDescription, short_name
 )
-from mypyc.rtypes import RType,  bool_rprimitive
+from mypyc.ir.rtypes import RType,  bool_rprimitive
 
 
 # Primitive binary ops (key is operator such as '+')
